Revert "Merged Manjikian Wednesday changes with clara's changes"
This reverts commit 37efaa5916d4c6120ed72f440ddc2f65e218299b.
