+++ qcore.h
