+++ qinit.c
