+++ qinit.h
