+++ querk.bat
